+++ src/App.tsx
@@ -1,7 +1,7 @@
 import './App.css';
 import React, { useContext } from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import { AppContext } from './index';
+import { AppContext } from './context/AppContext';
 import IAppContext from './interfaces/IAppContext';
 import Header from './components/Header/Header';
 import Home from './components/Home/Home';
