+++ src/context/AppContext.ts
@@ -1,0 +1,4 @@
+import React from 'react';
+import IAppContext from '../interfaces/IAppContext';
+
+export const AppContext = React.createContext<IAppContext>({} as IAppContext);
